+++ builder_pkgs/mockito/test/builder/auto_mocks_test.dart
@@ -3139,7 +3139,7 @@ void main() {
         class _Foo {}
         '''),
       },
-      message: contains('Mockito cannot mock a private type: _Foo.'),
+      message: contains("Mockito cannot mock a private type: '_Foo'."),
     );
   });
 
@@ -3266,7 +3266,7 @@ void main() {
         enum Foo {}
         '''),
       },
-      message: 'Mockito cannot mock an enum: Foo',
+      message: "Mockito cannot mock an enum: 'Foo'",
     );
   });
 
@@ -3293,7 +3293,73 @@ void main() {
         void main() {}
         '''),
       },
-      message: contains('Mockito cannot mock a non-subtypable type: int'),
+      message: contains("Mockito cannot mock a non-subtypable type: 'int'"),
+    );
+  });
+
+  test('throws when GenerateMocks references a sealed class', () async {
+    _expectBuilderThrows(
+      assets: {
+        ...annotationsAsset,
+        'foo|test/foo_test.dart': dedent('''
+        // @dart=3.0
+        import 'package:mockito/annotations.dart';
+        sealed class Foo {}
+        @GenerateMocks([Foo])
+        void main() {}
+        '''),
+      },
+      message: contains("Mockito cannot mock a sealed class 'Foo'"),
+    );
+  });
+
+  test('throws when GenerateMocks references sealed a class via typedef',
+      () async {
+    _expectBuilderThrows(
+      assets: {
+        ...annotationsAsset,
+        'foo|test/foo_test.dart': dedent('''
+        // @dart=3.0
+        import 'package:mockito/annotations.dart';
+        sealed class Foo {}
+        typedef Bar = Foo;
+        @GenerateMocks([Bar])
+        void main() {}
+        '''),
+      },
+      message: contains("Mockito cannot mock a sealed class 'Foo'"),
+    );
+  });
+
+  test('throws when GenerateMocks references a base class', () async {
+    _expectBuilderThrows(
+      assets: {
+        ...annotationsAsset,
+        'foo|test/foo_test.dart': dedent('''
+        // @dart=3.0
+        import 'package:mockito/annotations.dart';
+        base class Foo {}
+        @GenerateMocks([Foo])
+        void main() {}
+        '''),
+      },
+      message: contains("Mockito cannot mock a base class 'Foo'"),
+    );
+  });
+
+  test('throws when GenerateMocks references a final class', () async {
+    _expectBuilderThrows(
+      assets: {
+        ...annotationsAsset,
+        'foo|test/foo_test.dart': dedent('''
+        // @dart=3.0
+        import 'package:mockito/annotations.dart';
+        final class Foo {}
+        @GenerateMocks([Foo])
+        void main() {}
+        '''),
+      },
+      message: contains("Mockito cannot mock a final class 'Foo'"),
     );
   });